chore(tools): first draft splitting metabase export functions (#4825)
